Reorganize project into packages
