+++ proxy-robots-module/src/main/java/orwell/proxy/robot/Camera.java
@@ -1,4 +1,4 @@
-package orwell.proxy;
+package orwell.proxy.robot;
 
 public class Camera implements ICamera {
