+++ proxy-robots-module/src/main/java/orwell/proxy/robot/ColourWrapper.java
@@ -1,4 +1,4 @@
-package orwell.proxy;
+package orwell.proxy.robot;
 
 import orwell.messages.Robot;
