+++ proxy-robots-module/src/main/java/orwell/proxy/robot/RobotsMap.java
@@ -1,7 +1,8 @@
-package orwell.proxy;
+package orwell.proxy.robot;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import orwell.proxy.zmq.ZmqMessageDecoder;
 
 import java.util.ArrayList;
