+++ proxy-robots-module/src/test/java/orwell/proxy/ProtobufTest.java
@@ -7,6 +7,7 @@ import org.junit.runners.JUnit4;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import orwell.messages.Controller;
+import orwell.proxy.robot.Tank;
 
 import static org.junit.Assert.*;
